requirements.txt
Flask==2.0.1
pywebview==3.6.3

